Move State from converter.py to state.py
